backend/pyproject.toml
[tool.poetry]
name = "bookworm"
version = "0.1.0"
description = ""
authors = ["TBlackford <TBlackford@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
psycopg2 = "^2.9.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
uuid = "^1.30"
django-cors-headers = "^3.10.1"
djangorestframework-camel-case = "^1.3.0"
python-dotenv = "^0.19.2"
django-taggit = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
